--- styles/feedback.module.css
@@ -1,113 +1,0 @@
-body {
-    background-color: rgb(247, 245, 245);
-}
-.feedback {
-    margin: 5% 12%;
-    border-radius: 4%;
-    background-color: white;
-    box-shadow: 0px 0px 2px rgb(197, 196, 196);
-    border: 0px;
-}
-
-.slide {
-    flex-direction: row;
-    display: flex;
-    -webkit-animation-name: fade;
-    -webkit-animation-duration: 1.5s;
-    animation-name: fade;
-    animation-duration: 1.5s;
-}
-@-webkit-keyframes fade {
-    from {opacity: .4}
-    to {opacity: 1}
-  }
-  
-  @keyframes fade {
-    from {opacity: .4}
-    to {opacity: 1}
-  }
-
-.content {
-    margin: 5%;
-    padding-top: 5%;
-
-}
-
-.forImage {
-    display: inline-block;
-   
-}
-
-img {
-    border-radius: 15%;
-    width: 600px;
-}
-
-.prev,.next {
-    cursor: pointer;
-    position: relative;
-    width: auto;
-    padding: 16px;
-    margin-top: -22px;
-    color: white;
-    font-weight: bold;
-    font-size: 18px;
-    transition: 0.6s ease;
-    border-radius: 0 3px 3px 0;
-    user-select: none;
-}
-
-.next {
-    right: 0;
-    margin: auto;
-    border-radius: 3px 0 0 3px;
-}
-
-.prev {
-    left: 0;
-    margin: auto;
-    border-radius: 3px 0 0 3px;
-}
-
-.prev:hover,.next:hover {
-    background-color: rgba(0, 0, 0, 0.8);
-}
-
-h1 {
-    font-family: 'Readex Pro', sans-serif;
-    font-weight: 400;
-    font-size: 35px;
-    margin-bottom: 5px;
-}
-
-h2 {
-    font-family: 'Poppins', sans-serif;
-    font-weight: 300;
-    font-size: 20px;
-    color: rgb(168, 168, 168);
-    margin: 0%;
-}
-
-.rating {
-    margin-top: 5px;
-    margin-bottom: 20px;
-    color: yellow;
-    
-}
-
-.review {
-    font-family: 'Poppins', sans-serif;
-    font-weight: 300;
-    font-size: 17px;
-    color: rgb(168, 168, 168);
-}
-
-@media only screen and (max-width:500px)
-{
-    .slide{
-        flex-direction: column;
-    }
-    img{
-        width: 300px;
-    }
-}
